feat(YouTube Music): Add `Remove viewer discretion dialog` patch
