+++ patches/src/main/kotlin/app/morphe/patches/music/interaction/dialog/RemoveViewerDiscretionDialogPatch.kt
@@ -5,23 +5,23 @@
  * See the included NOTICE file for GPLv3 Section 7 terms that apply to this code.
  */
 
-package app.morphe.patches.music.misc.quic
+package app.morphe.patches.music.interaction.dialog
 
 import app.morphe.patches.music.misc.extension.sharedExtensionPatch
 import app.morphe.patches.music.misc.settings.PreferenceScreen
 import app.morphe.patches.music.misc.settings.settingsPatch
 import app.morphe.patches.music.shared.Constants.COMPATIBILITY_YOUTUBE_MUSIC
-import app.morphe.patches.shared.misc.quic.disableQUICProtocolPatch
+import app.morphe.patches.shared.interaction.dialog.removeViewerDiscretionDialogPatch
 
 @Suppress("unused")
-val disableQUICProtocolPatchMusic = disableQUICProtocolPatch(
+val removeViewerDiscretionDialogPatch = removeViewerDiscretionDialogPatch(
     block = {
         dependsOn(
             sharedExtensionPatch,
-            settingsPatch
+            settingsPatch,
         )
 
         compatibleWith(COMPATIBILITY_YOUTUBE_MUSIC)
     },
-    preferenceScreen = PreferenceScreen.MISC
+    preferenceScreen = PreferenceScreen.GENERAL
 )
