+++ patches/src/main/kotlin/app/morphe/patches/shared/interaction/dialog/Fingerprints.kt
@@ -1,4 +1,11 @@
-package app.morphe.patches.youtube.interaction.dialog
+/*
+ * Copyright 2026 Morphe.
+ * https://github.com/MorpheApp/morphe-patches
+ *
+ * See the included NOTICE file for GPLv3 Section 7 terms that apply to this code.
+ */
+
+package app.morphe.patches.shared.interaction.dialog
 
 import app.morphe.patcher.Fingerprint
 import app.morphe.patcher.InstructionLocation.MatchAfterImmediately
@@ -42,4 +49,3 @@ internal object AdultContentSetPropertiesFingerprint : Fingerprint(
         string("allowControversialContent", location = MatchAfterImmediately()),
     )
 )
-
