+++ patches/src/main/kotlin/app/morphe/patches/youtube/misc/quic/DisableQUICProtocolPatch.kt
@@ -1,3 +1,10 @@
+/*
+ * Copyright 2026 Morphe.
+ * https://github.com/MorpheApp/morphe-patches
+ *
+ * See the included NOTICE file for GPLv3 Section 7 terms that apply to this code.
+ */
+
 package app.morphe.patches.youtube.misc.quic
 
 import app.morphe.patches.shared.misc.quic.disableQUICProtocolPatch
