+++ patches/src/main/kotlin/app/morphe/patches/youtube/video/audio/DisableDRCAudioPatch.kt
@@ -1,3 +1,10 @@
+/*
+ * Copyright 2026 Morphe.
+ * https://github.com/MorpheApp/morphe-patches
+ *
+ * See the included NOTICE file for GPLv3 Section 7 terms that apply to this code.
+ */
+
 package app.morphe.patches.youtube.video.audio
 
 import app.morphe.patches.shared.misc.audio.drc.disableDRCAudioPatch
