+++ patches/src/main/resources/addresources/values/shared-youtube/strings.xml
@@ -177,6 +177,10 @@ However, enabling this will also log some user data such as your IP address."</s
     <string name="morphe_settings_menu_filter_screen_summary">Hide items on the standard app settings screen by their visible name</string>
     <string name="morphe_settings_menu_filter_picker_empty">Open the standard app settings screen once to populate the list</string>
 
+    <!-- interaction.dialog.removeViewerDiscretionDialogPatch -->
+    <string name="morphe_remove_viewer_discretion_dialog_title">Remove viewer discretion dialog</string>
+    <string name="morphe_remove_viewer_discretion_dialog_summary">This does not bypass the age restriction. It just accepts it automatically</string>
+
     <!-- misc.spoof.appversion.spoofAppVersionPatch -->
     <string name="morphe_spoof_app_version_title">Spoof app version</string>
     <string name="morphe_spoof_app_version_summary">Tricks the server into thinking you are using an older app version to restore legacy UI features</string>
