+++ patches/src/main/resources/addresources/values/youtube/strings.xml
@@ -357,10 +357,6 @@ Tap here to open AiSlopList.com and see downloads for other platforms"</string>
     <string name="morphe_copy_video_link_with_timestamp_button_title">Copy timestamp by default</string>
     <string name="morphe_copy_video_link_with_timestamp_button_summary">Reverses the tap and tap-and-hold actions of \'Copy video link\' button</string>
 
-    <!-- interaction.dialog.removeViewerDiscretionDialogPatch -->
-    <string name="morphe_remove_viewer_discretion_dialog_title">Remove viewer discretion dialog</string>
-    <string name="morphe_remove_viewer_discretion_dialog_summary">This does not bypass the age restriction. It just accepts it automatically</string>
-
     <!-- layout.hide.signintotv.disableSignInToTVPopupPatch -->
     <string name="morphe_disable_sign_in_to_tv_popup_title">Disable \'Sign in to TV\' popup</string>
